Do all settings updates from the main process
Instead of making writing the settings directly from the content process,
send messages back to the main process and write things out there.  This
ensures that the main process has a matching view of the settings at all
times, and that the two sides will not get out of sync.