vue.config.js: skip babel in node_modules
There is no reason to run babel for things in node_modules (which should
already be compiled); add an exclude rule for it.  This reduces the compile
time from ~50s down to ~10s; this is not quite as good as not running babel at
all (which was at ~8s), but should already be a significant improvement.