Downgrade to Vue 2
We want to be pulling things in from Rancher Dashboard, but that's still using
Vue 2.  In order to encourage code sharing, we'll (temporarily) downgrade to
Vue 2 to match to make pulling in changes easier.  We will eventually move
back up to Vue 3, of course, but that will be longer term.