Pages: rename index back to Welcome
Using .nuxt-link-exact-active will not work as we extend our pages,
since they may pick up additional parameters (as query strings), which
will cause the selectors to no longer match.  Instead, add a middleware
that will redirect the index page to the Welcome page instead, so that
we can return to using .nuxt-link-active.