k8s-engine: Backend support for multiple port forwards
This refactors the code such that we can handle multiple port
forwardings (though currently no code makes use of such).  Also, move
the code to probe the port forwarding to the homestead part, as
intentionall ignoring invalid certificates and expecting a passing HTTP
response are too specific for general use.

Signed-off-by: Mark Yen <mark.yen@suse.com>