Add dev script
Nuxtron works more like a dev server, in that it watches for changes in
the main process and restarts everything on change.  This does not work
well for us, as that means we keep restarting minikube in the background
(which takes a significant amount of time).  Instead, switch to our own
dev script which mostly copies from nuxtron except for not watching in
the background.