ESLint: Disable rules not actve in dashboard
These rules don't appear to be active in dashboard's linting.  Disable
them manually to avoid them triggering unexpectedly and changing
dashboard code.

I suspect they are active for us because of different versions of
various eslint plugins being in use; however, I could not find a
combination that will work with the versions from the dashboard.

Signed-off-by: Mark Yen <mark.yen@suse.com>