scripts/dev: allow waiting for renderer to be ready
When running the renderer process, nuxt can end up earsing some output
as it builds things.  If we are running serially (and therefore likely
to care about the output), wait until the welcome page is ready before
continuing, taking that as an indicator that nuxt has stablized.

Using the welcome page is necessary as nuxt doesn't seem to otherwise
have a way of determining when the initial building is complete.

Signed-off-by: Mark Yen <mark.yen@suse.com>