Removing logs that are showing as error
minikube is putting out debug/log info on stderr. This is a
common practice. Even when the exit code is 0 (as in everything
is fine) we had been rejecting if there was any data from stderr.
Logs were causing this. Removing the rejection if there is
data in stderr but it exited properly.

Signed-off-by: Matt Farina <matt.farina@suse.com>