Create setup script for Windows
This converts setupmac.js into a generic download-resources.mjs and make it
work on Windows.  This also drop the follow-redirects dependency in favour
of using node-fetch (which we already need in the app).  Additionally,
reduce the callback nesting we have in the same script.

There doesn't seem to be significant reason to convert the script to async
(considering that this is run as a one-off).

Additionally, make all this get called as a postinstall hook, so that it
happens automatically upon `npm install`, and the user no longer needs to
worry about it.

Signed-off-by: Mark Yen <mark.yen@suse.com>