scripts/windows-setup: Work around Python deprecation
Python 2.7 is long obsolete, but node-gyp still wants it.  This means that
attempting to install it displays a deprecation warning; PowerShell will
see this warning (on stderr) and assume it means that the command
failed.  Work around this by redirecting stderr for that command, and
instead letting exit statuses control failure detection.

Signed-off-by: Mark Yen <mark.yen@suse.com>