Jest: Ignore build & git directories
These directories can never contain node modules we care about for tests.
Jest was keeping a watch on each directory, which causes issues on Windows
as opened directories cannot be moved (which breaks the build process if
there is a background jest process running).  The git directory part is
purely optimization.

Signed-off-by: Mark Yen <mark.yen@suse.com>