Improve robustness
Sometimes, on windows VMs, `[component/Images].imageOutputCuller`
becomes null unexpectedly. Allow for this so the `close output text'
button appears, and other UX happens smoothly.

No idea why this happens. I added `console.log` statements to track
when this field was assigned and didn't see an assigment from non-null
to null, but on event-end it was `null`.

Signed-off-by: Eric Promislow <epromislow@suse.com>