Add helper for spawning child processes.
We repeatedly write code to handle spawning children that log to a file and
return a promise.  This attempts to factor that out.

Note that on Windows we may need to to string encoding conversion before
writing the output, as some tools may emit UTF-16 instead.  This has built-
in support for that.

Signed-off-by: Mark Yen <mark.yen@suse.com>