Upgrade to Electron 12 (currently 12.0.9)
We were affected by one breaking change:
https://www.electronjs.org/docs/breaking-changes#default-changed-contextisolation-defaults-to-true:

'''
In Electron 12, contextIsolation will be enabled by default. To restore the previous behavior, contextIsolation: false must be specified in WebPreferences.
'''

This compromises security but enables `window.require`. Some of the external processes
we spawn take user-supplied arguments. As long as we run them using array-based methods like
`child_process.execFile(Sync)` and `child_process.spawn(Sync)`,
and not `child_process.exec(Sync)` we should be ok.

Also pull the `@types/node` entry out of `devDependencies` as electron pulls it in
and we don't directly depend on it.

Signed-off-by: Eric Promislow <epromislow@suse.com>