Explain why a link doesn't hold in the checkbox label
Why: because dashboard tooltips present as a small unlabeled button,
clicking on which displays the underlying text in a hover tip.

We've already seen the dashboard tooltip at work and rejected it.

The backend is responsible for figuring out why a link can't happen,
and the frontend displays it with a minimum of processing.

Signed-off-by: Eric Promislow <epromislow@suse.com>