Incrementing cli tool versions
Helm, kubectl, and kim all have newer releases. This commit updates
to newer versions.

Note, kubectl is not on the latest minor release as to have a
larger support window of versions and because the latest k8s minor
is only 1.20.1. It's useful to wait for a further patch version to
be released before upgrading.

Signed-off-by: Matt Farina <matt.farina@suse.com>