When stopping hyperkit, don't create a promise
When hyperkit is stopped inside a promise, the shutdown mechanism
in Electron issues an unresolved-promise warning just before the
promise is actually resolved. The simplest solution is to replace
the promise associated with the outcome of running `hyperkit stop`
with a trivial promise that resolves earlier, as we don't care about
the outcome of stopping the server (because we're shutting down).

Signed-off-by: Eric Promislow <epromislow@suse.com>