Hyperkit: Check valid cert when waiting for K3s to be up.
When we initially start, K3s could have an incorrect certificate if it had
previous ran but the VM had changed its IP address in the mean time (as the
certificate would have the old address but not the new one).  This leads to
errors when trying to fetch the kubeconfig.  Guard for that by explicitly
checking for the correct certificate before continuing.

This is in shared code as a similar thing needs to be done in the WSL
backend.

Signed-off-by: Mark Yen <mark.yen@suse.com>