K3s: Simplify getting the kubeconfig
Since we need to buffer the whole kubeconfig in memory for the parser
anyway, it's easier to just run the command and capture the output rather
than trying to write it to disk for no reason.  That means it's simpler to
get the caller to return the configs directly, rather than trying to go
through some useless indirection.

We were not attempting to retry getting the kubeconfig when it fails; don't
pretend we'd do so in the logs.

Signed-off-by: Mark Yen <mark.yen@suse.com>