Building on tags
The internal version that's displayed in the UI comes from the
git tag. This means we need builds associated with the tag to
display properly and cannot use the last run on main that the
tag is associated with.

This change causes builds to be run on tags as well as the main
branch and pull requests.

Signed-off-by: Matt Farina <matt.farina@suse.com>