Pages/kim: Check if kim is ready.
With the new kim state from the last commit, we can propagate that to the
UI now so that we have more granular state on the images page, to let the
user know that although the cluster is ready, the image manager isn't yet.
This should reduce confusion when the user sees that the images list is
blank when they don't expect it to be.

Signed-off-by: Mark Yen <mark.yen@suse.com>