Electron-builder: Create custom publisher
We plan on using Longhorn's upgrade-responder to handle upgrade pings,
backed by GitHub releases for the actual binaries.  As a first step, we
need to create a custom electron-builder Publisher backend to generate the
necessary fields to be examined durin the upgrade ping process.

Note that this publisher does not actually publish anything; it just
resolves the options at build time to be written into the configuration
file.

Signed-off-by: Mark Yen <mark.yen@suse.com>