Update: Implement re-checking and minimum interval
We should not always check for updates on startup; rather, we need to
record when the next update check is expected, and only run the check if
it's at least that time.  Also, we need to check for updates again after
some amount of uptime, depending on what the server response says.

Signed-off-by: Mark Yen <mark.yen@suse.com>