WSL: Ignore expected error removing log files during factory reset.
We open log file streams (with FILE_SHARE_DELETE per NodeJS defaults)
during the lifetime of the app.  This means that attempting to delete the
directory will result in ENOTEMPTY errors, as the _directory_ holding the
log files will be locked and cannot be removed.  Catch this during factory
reset and drop it on the floor, as the files of interest are removed.

Signed-off-by: Mark Yen <mark.yen@suse.com>