Updater: Cache status internally in provider.
This lets us always trigger update requests; the provider will do any
caching required to obey the update interval requests (which ensures that
we don't dirty the stats).  This ensures we can apply updates on start
correctly, as electron-updater requires an update check to set up internal
state before we can apply updates.

We also move things around in the main process so that we can exit earlier
to apply updates as necessary.

Signed-off-by: Mark Yen <mark.yen@suse.com>