scripts/sign: Only build nsis when signing
By default, we would end up building all targets in parallel.  This would
cause issues as the zip build would run 7zip to create the archive, which
would lock the executables, breaking the signing (or, alternatively, the
signing would block the zip creation, depending on a race).  This can be
fixed by only building the installer when signing instead, as we will not
be using the zip artifact from this build.

Signed-off-by: Mark Yen <mark.yen@suse.com>