Remove hyperkit files when moving to lima
This only happens when:
1. running on macos
2. Starting up with a preferences file of version 1 or 2

This means in the future we can safely use the directory
~/Library/State/rancher-desktop/driver
as long as the settings file has a version # > 3

Yes, this is piggy-backing on the settings upgrade mechanism
to avoid writing yet another one.  If you think of it as
an `at-startup-upgrade mechanism' it's less problematic.

Signed-off-by: Eric Promislow <epromislow@suse.com>