Lima: Force-terminate any stray k3s instances before start
When we restart k3s (e.g. when we change versions), we can end up with a
stray k3s process on a VM.  This often does not come with a matching host-
side porocess.  Force terminate any k3s instances before we start a new one
to ensure that we can listen on the port correctly.

Signed-off-by: Mark Yen <mark.yen@suse.com>