K3s: Try to be more explicit about where we want the short version.
Since k3s internally wants a version string with the k3s build number (as
in v1.2.3+k3s.4), we need to do a better job of tracking which we mean when
we refer to a version.  This is unfortunately advisory only; TypeScript
does not see them as distinct types, and does no type checking.

We do change to making the desired version be the short version, as we will
persist it if we successfully start without a version saved.

Signed-off-by: Mark Yen <mark.yen@suse.com>