Lima/WSL: Abort if user stopped before start finished
If the user attempts to quit before start has finished (e.g. while waiting
for the k3s components to download), quit immediately, instead of
continuing to start.  Otherwise the quit flag would be set without ever
attempting to quit, causing the app to orphan an instance if the user tries
again later.

Signed-off-by: Mark Yen <mark.yen@suse.com>