Dev only: restore the cache file (or absence of) after tests
This was manifesting as a failure to fetch
https://github.com/k3s-io/k3s/releases/download/v1.20.0/k3s-airgap-images-amd64.tar
at startup, and rd would freeze.

`~/Library/Caches/rancher-desktop/k3s-versions.json` contained one entry:

`["v1.20.0"]`

This is because unit tests created a cache file with that entry.

A partial fix is to restore the existing list of versions after running
tests, done here.

A better fix would be to use a different cache file, but that's probably
overkill for something that affects only devs. This doesn't affect the
CI system (which runs tests but not the final build) or end-users
(who run final builds but not unit tests).

Signed-off-by: Eric Promislow <epromislow@suse.com>