Implement file shadowing detection
Currently, these are `kim`, `kubectl`, and `kuberlr`

- Report if there's an older version of that file earlier in the path
- Report if there's a newer version of that file later in the path
- For `kubectl`, report any instances elsewhere in the path, because
  our kubectl is actually a link to kuberlr, which manages versions for us.

This change has RD watch for changes in each directory in the path.
When there's a change involving one of the files we care about, it refreshes
the shadow-info.

When the preferences window is shown, it should use current info.  So all
the directories in the path are walked only at the start, and when there's
a change in one of them.

Signed-off-by: Eric Promislow <epromislow@suse.com>