Lima: Use openrc to manage k3s.
Long-running instances can encounter an issue where the ssh session to the
VM is interrupted even though nothing else is wrong; this leads to the
handle to the k3s child process disconnecting, which causes the VM to go
into an error state.  Instead, use openrc to manage the lifetime of the k3s
process (restarting it as necessary), and watch the logs with tail.

Signed-off-by: Mark Yen <mark.yen@suse.com>