Background: don't open preference before protocol registration.
It appears that it is possible to click on the dock icon before the app is
ready, in which case we attempt to open the preferences window too early.
This reports an error saying `Cannot create BrowserWindow before app is
ready`.

Instead, we force wait until the protocol handler has been registered
(after the app is ready); we delay it this far as otherwise attempting to
open the window leads to an unknown protocol error and a white window.

Signed-off-by: Mark Yen <mark.yen@suse.com>