k3s: Wait for cert to be ready.
When a cert is freshly generated (e.g. because the WSL distribution doesn't
exist yet), the cert might not be ready by the time we examine it.  We need
to manually wait for it to be ready, to ensure that it doesn't get seen by
other processes (e.g. `kubectl version`).

Signed-off-by: Mark Yen <mark.yen@suse.com>