WSL: Do argument translation for nerdctl
This updates nerdctl-stub to parse arguments (such as
`nerdctl build --file`) and translate the paths.  On Windows (the host),
this just means convert it to `/mnt/c/...` form; on Linux (a second WSL
distribution), this means making a temporary bind mount under `/mnt/wsl/`
and converting relevant arguments to point to that instead.

In order to do the mounting, the nerdctl-stub binary needs to be setuid
root.  We therefore need to copy it to somewhere under `/mnt/wsl/` as well,
since anything under `/mnt/c/` (DrvFS) doesn't support setuid properly.

In order to correctly parse the arguments, we use a generator that parses
the output of `nerdctl -help` (and various subcommands) to build our
understanding of the available subcommands and options.  This may be
triggered on a Windows machine with the `rancher-desktop` WSL distribution
using `npm run generate:nerdctl-stub`.  (The requirement is to be able to
execute the `nerdctl` binary to parse its output.)

Signed-off-by: Mark Yen <mark.yen@suse.com>