Remove code causing node roles to be deleted
At one point we renamed the instance from lima-rancher-desktop to
lima-0. The rancher-desktop and 0 are names passed to lima. The
full name shows up in the socket path which has a limited number
of characters. The shorter the name the better.

Recently, we set the host name back to rancher-desktop. This is
internal to the OS running on the node and not lima. This caused
the node name to again be lima-rancher-desktop.

The code deleted in this change removes the cleanup. The cleanup
was deleting the node. It was then being added back by k3s without
any roles on it.

Closes #667

Signed-off-by: Matt Farina <matt.farina@suse.com>