Restoring the integrations location on Windows
Windows does not need this location but there are code paths that
do not handle an error being returned. There are multiple code paths
like this. Restoring the mac location, which has long been used.

In a future change, this can be removed and the other code paths fixed
to handle an error being returned.

Signed-off-by: Matt Farina <matt.farina@suse.com>