Updates: don't attempt to apply updates if it's not newer.
When we check for updates, we set an `isInstallable` flag in our cache if
the new version has been downloaded.  However, that flag isn't being
cleared right after an update (because Squirrel.Mac etc. doesn't know about
it).  Sanity check that flag by comparing the app version against the
staged app version, to make sure that we only abort startup for updates
that we may actually apply.

Signed-off-by: Mark Yen <mark.yen@suse.com>