WSL-helper: implement basic docker socket proxying.
This implements docker socket proxying for both Windows and Linux for the
default configuration.  The `wsl-helper docker-proxy start` subcommand
starts dockerd.

For Linux, we make docker listen on a socket in /mnt/wsl/ so that it is
exposed to all distributions; we then have `wsl-helper docker-proxy serve`
which forwards everything from the default docker endpoint
(`/var/run/docker.sock`) to it.

For Windows, we additionally make `wsl-helper docker-proxy start` listen on
AF_VSOCK; the Windows-side `wsl-helper docker-proxy serve` can then connect
to it (by using Hyper-V APIs) and do the forwarding.

We will need further logic in the forwarder (the `serve` part) to handle
path translations in the future.

Signed-off-by: Mark Yen <mark.yen@suse.com>