Launch k3s with the specified `--flannel-iface NAME` option
Note that when the option isn't specified, the network name is "lima0",
and the VM needs to be rebuilt with the new interface name.

I pulled out code that ran `ip a` and `ps auxww | grep k3s.*server` remotely
to verify that the running networks aligned. I found that they would
align with a new value in the config file, but ingress didn't work
until the VM was rebuilt. So this might need documentation, or further exploring.

Also, the config was being calculated by merging the default config over the
user's current settings. This only works if we want to give the default settings
priority over any changes the user might wish to make.

Signed-off-by: Eric Promislow <epromislow@suse.com>