WSL: Use OpenRC.
This stops tracking the k3s process directly; instead, we run OpenRC within
the WSL distribution.  This lets us run mutiple processes within the same
mount namespace, which will be needed for moby.

This also lets us use the machine-id service instead of generating it
manually, to closer match the macOS path.  We also create a service for the
guest agent, so we no longer need to keep track of its process.

Signed-off-by: Mark Yen <mark.yen@suse.com>