WSL: make subprocesses log to different files.
It appears that when multiple processes log to the same file, they can step
on each other and corrupt the log file (or, when it's working, overwrite
each other's log lines).  Use different log files for different things to
help reduce the chances of this happening, until we can fix the logging
later to actually fix the issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>