WSL: Stop wsl-helper manually.
It appears that killing the owning wsl.exe process doesn't always stop the
child process.  Add a new `wsl-helper docker-proxy kill` command to kill
any server processes.  As the command name is now different, we need to
update how we find the server process to ensure we kill the correct
processes.

Also ensure that we write to the distro-specific log file.

Signed-off-by: Mark Yen <mark.yen@suse.com>