WSL-Helper: implement helper for factory reset.
On Windows, we're not allowed to delete files which are in use.  However,
that means we'll never be able to remove files in use by Electron (such as
the Chrome user profile directory).

This implements a helper subcommand in wsl-helper that can be used to
delete the user data, so that the new flow would be:

- Spawn `wsl-helper factory-reset ...`
- Quit the Electron processes
- The files get deleted
- Respawn Rancher Desktop

Note that this is not yet used by the UI.

Signed-off-by: Mark Yen <mark.yen@suse.com>