OpenRC: quote log files.
The script that OpenRC uses to launch supervise-daemon doesn't correctly
quote the variables; this means that it will get passed unquoted, so if the
log path contains spaces it will be split as separate arguments and the
second part (and later) will be taken as the command.  Fix that by quoting
things manually.

Note that this means when we use the variables internally (e.g. passing to
checkpath), we need to use the correct quoting.

See also:
https://github.com/OpenRC/openrc/blob/0.44.9/sh/supervise-daemon.sh

Signed-off-by: Mark Yen <mark.yen@suse.com>