Use vmnet interface only for ingress
Throughput is almost an order of magnitude lower on the vmnet
interface then via slirp. It is unknown if this is related to
using libvdeplug, or to vde_vmnet itself.

This commit also changes the network mode to "bridged" so that
the routable IP address also works outside the host itself.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>