ensure that cirrus will automatically triggers tasks when invoked by cron
On Cirrus CI, a CI job configured for manual triggering isn't triggered
even if it matches a cron expression set in the repository options. By
making it trigger_type: automatic when invoked via cron, we can
ensure the correct thing happens, and that the jobs are triggered
either via cron or via a manual trigger (but not on push).

Signed-off-by: evertonlperes <evertonluperes@gmail.com>