Updater: Clearer code flow for setting up updater.
This turns the update setup into a state machine, with a clear difference
between having configured the updater and having used it to check for
updates.  This ensures that we don't accidentally create two instances of
the updater (one from loading settings, and the the one in the startup path
we expect).

Signed-off-by: Mark Yen <mark.yen@suse.com>