Removed unnecessary tmp directory delete
This is related to the issue: https://github.com/rancher-sandbox/rancher-desktop/issues/1165 - We were trying to remove a tmp folder that was renamed before being deleted. In this case, the function safeRename(), will rename the tmp folder and the tmp folder no longer exists in the system, not being necessary for removing an unexisting tmp folder.

Signed-off-by: Everton Peres <37411123+evertonlperes@users.noreply.github.com>