Movifing artifacts upload due a bug from cirrus ci runner
After cirrus changes, it still occuring the error if you try to upload a file outside of the runner main folder (/tmp/cirrus-ci-build/). Needed to refactor a bit in order to move the artifacts that I need from the non-user project folder to root project folder. Filing a new issue to Cirrus CI in order to fix it.

Signed-off-by: Everton Peres <37411123+evertonlperes@users.noreply.github.com>