Added one more waitFor before start the background services
Just realize the progress bar takes roughly ~60s to be visible, before start the background services. I've added another waitFor in order to check if the progress bar will be visible and than wait util the progress bar be detached (hidden). This is a workaround until we find a better way to make sure all backend services are up and running.

Signed-off-by: Everton Peres <37411123+evertonlperes@users.noreply.github.com>