WSL: Customize /etc/hosts
This edits /etc/hosts on the main distribution to add entries for
host.docker.internal and host.minikube.internal for compatibility.  We copy
over the /etc/hosts file in the data distribution (which we still ask WSL
to generate), and then add lines we need.

Note that this doesn't appear to actually affect any containers, as moby
uses an internal file to generate /etc/hosts.

Signed-off-by: Mark Yen <mark.yen@suse.com>