Restore port forwaring on mac/linux
The change to remove the port forwarding was part of the feature
to automatically forward ports from the VM to the host. We
realized on the Windows side that forwarding internal services
(e.g., the database for a site that's not exposed) should be
available all the time. So, this wasn't removed from Windows. It
should have been restored to mac and linux earlier. Noticed while
doing docs.

Signed-off-by: Matt Farina <matt.farina@suse.com>