wsl-helper: close /attach on /wait exit.
On Windows, running interactive docker sessions does not quit correctly
when the container exits; instead, the user must press a key to let the
client notice the end of the connection.  This does not happen on Linux (in
WSL).  The underlying API (/containers/{id}/wait) appears to work correctly
however; the connection closes correctly.  As a workaround, forcefully
close any /containers/{id}/attach connections when a corresponding /wait
returns.

Signed-off-by: Mark Yen <mark.yen@suse.com>