WSL-Helper: make Linux vsock connection non-blocking.
When we pipe the connection between the unix docker socket
(/var/run/docker.sock) and the vsock socket, we need to open the vsock side
as non-blocking so that once the unix side is closed we can close the vsock
side properly.  Otherwise the client would not be able to detect the socket
should close.

This copies an (Apache-2.0 licensed) file from github.com/linuxkit/virtsock
so that we can make the resulting socket non-blocking.  The only relevant
change is in `func (v *vsockListener) Accept()` to add `SOCK_NONBLOCK`.

Signed-off-by: Mark Yen <mark.yen@suse.com>
(cherry picked from commit 10e29e0e69589afa22cf9945b125e3234ca512fe)