Use a more specific test for the presence of a file or symlink.
Unforunately, in the ubuntu/wsl subsystem, `-e GOOD_SYMLINK` fails.
So I need to test both `-e` and `-L` to verify there's no file or symlink
with the given name present.

Note that `-f` and `-e` work the same on macos and linux, but this is wsl-only code.
We need to test for both any file (the `-e`) and symlinks (the `-L`).

Signed-off-by: Eric Promislow <epromislow@suse.com>