Refactored navigate to, minor improvements on main spec and improved pages
TypeScript seems to have issues determining the correct type if we just use
a constructor as the return type; instead, we have to use an inline
function to convince it to derive the correct (class) type as a return
value.  Additionally, declaring the type of navigateTo() on a separate line
bypasses issues where it isn't quite smart enough to determine that the
value we _do_ return is correct, letting us skip a cast. by @mook-as

Signed-off-by: Everton Peres <37411123+evertonlperes@users.noreply.github.com>