fetch: wrap existing agent instead of new subclass.
This ensures that we can continue to use the existing agent for other
functionality (in particular, to handle the proxies).  This means we no
longer need to explicitly set the CA certificates (since we no longer
construct new instances of https.Agent).

This also includes a fix to ElectronProxyAgent to call
agent.createConnection correctly (to work with the fetch wrapper).

Signed-off-by: Mark Yen <mark.yen@suse.com>