networking: Don't import expired CA certs.
We appear to have issues where people have expired CA certs on their
machine, leading to errors verifying otherwise valid server certificates
that can either chain to a valid (unexpired) CA or an expired CA.  Try to
avoid importing already expired CAs in hopes of fixing the issue.

It was not possible to use `forge.pki.certificateFromPem` directly as that
rejects non-RSA certificates (even though we do nothing with the public
key), so we have to copy some code over to be able to parse the certificate
manually from PEM just enough to see the validity period.

Signed-off-by: Mark Yen <mark.yen@suse.com>