Iptables rules without address range bind to 0.0.0.0
This commit bumps to lima 0.8.3+ to include the change that implements this.

It also bumps the wsl distro to 0.16, which in turn includes
rancher-desktop-agent v0.1.2, which uses the corresponding lima code.

The lima port forwarding rules are modified so that only ports bound to
0.0.0.0 are forwarded to 0.0.0.0 on the host. Ports bound to 127.0.0.1
are forwarded to 127.0.0.1 on the host, and all other interfaces are not
forwarded.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>