Do not relaunch when doing a factory reset
In order to use docker, the docker socket must exist. But in order to
create the docker socket at its standard location (/var/run/docker.sock)
we need root privileges. In order to execute code as root, we use the
sudo-prompt npm package. sudo-prompt uses pkexec on Linux. pkexec was
exiting with a code that indicated that it was not being started as
root. This of course is a problem because it must be run as root in
order to work. This is accomplished by setting the setuid bit on the
pkexec file and making its owner root. This was the case on all systems
that had the problem.

However, there is a case where the setuid bit is not respected: when the
no_new_privs bit is set on a process. As it turns out, this bit was set
for Rancher Desktop only after factory resetting and relaunching. It is
unclear why this is the case - a comment in the code for electron says:

// Chrome uses prctl(2) to set the NO_NEW_PRIVILEGES flag on Linux (see
//
https://github.com/torvalds/linux/blob/40fde647cc/Documentation/userspace-api/no_new_privs.rst).
// We disable this for unsandboxed processes, which the renderer tests
// are running in. If this test fails with an error like 'effective uid
// is not 0', then it's likely that our patch to prevent the flag from
// being set has become ineffective.

So it may be a bug in electron.

Other notes:
- You can see whether the no_new_privs bit is set for a process with `cat /proc/<pid>/status | grep NoNewPrivs`.
- This link is very helpful in understanding how this all works: https://ops.tips/notes/setuid-or-how-sudo-works/

Signed-off-by: Adam Pickering <adamkpickering@gmail.com>