Lima: refactor sudo-related code; no need to mutate instance.
This refactors the code to try to indicate that the sudo-related code does
not need to mutate the state of the LimaBackend instance.  Note that, due
to TypeScript limitations, Readonly<T> does not include the protected
fields of T, so we need to use `this: Readonly<this> & this` to include
them (but this means we don't get full protection from types).

With manual review, this should ensure that we can later support a mode in
which we have proper fallbacks for running without elevation.

Signed-off-by: Mark Yen <mark.yen@suse.com>