Add versioning to rdctl.
We're not doing much with the versions now,
but this is just to have *something* in place. Later the
CLI can ask the server what version it's running, and refuse
if there's a version mismatch. Or something. Right now I
just want the CLI to have a version.

And move "/v0" into a constant, so it's more obvious.

Signed-off-by: Eric Promislow <epromislow@suse.com>