Add value-correction to the server.
This wasn't needed when settings were changed only from the UI,
and it isn't needed when using our CLI. But because a wider
range of values can be submitted via the API, we need to convert
non-canonical values into their equivalent but canonical forms.

Do this in a separate routine, to get simpler code, and unit testing.

Signed-off-by: Eric Promislow <epromislow@suse.com>