WSL: Add service to generate dnsmasq configuration.
As WSL may regenerate the VM at any time no processes are running, we
cannot have the IP address in any configuration files before we run init.
As such, we switch to generate dnsmasq configuration (and /etc/resolv.conf)
in a service that runs before dnsmasq, to ensure that it is spawned via the
same init process.

Note that this is a one-shot service (that is, it exits quickly rather than
staying resident), so we need to use a custom start() function rather than
using the default that uses supervise-daemon.

Signed-off-by: Mark Yen <mark.yen@suse.com>