Use the statusMessage only if there's a failure to get the actual message.
When there's an error, the status message is generated by the http server,
and has a form like "400 Bad Request". If we fail to read the actual message
from the response.body, only then do we actually display the status message.

It's also displayed for unexpected error codes (not 400, 401, or 500), because
there's nothing from the command-server to display.

Signed-off-by: Eric Promislow <epromislow@suse.com>