Networking: Windows: use custom CA numeration
This replaces use of win-ca with a custom FFI-based CA enumeration.

On some systems with many CA certificates, the use of win-ca (which is
based on running a native executable and capturing stdout) causes issues
due to a failure to allocate enough buffer to capture the whole output.
Re-implement CA certificate enumeration using FFI to call Windows APIs
directly (which works as an iterator) to fix this.

As this iteration is (in theory) asynchronous, we need to add an event to
signify when this is complete (and we can make HTTPS requests).  This is
used by K3sHelper to ensure we fetch versions and download caches using the
system certificates.

Signed-off-by: Mark Yen <mark.yen@suse.com>