Remove legacy symlinks on Linux only
On macOS the situation is more complex. In the past
we have created symlinks in /usr/loca/bin/, which requires
sudo privileges to change. So, to delete legacy symlinks in
this directory we would need to check sudo privileges.
This doesn't fit with the way the code works at the moment,
so I'm leaving it for a future PR (if it matters enough,
that is).

Signed-off-by: Adam Pickering <adamkpickering@gmail.com>