Add Vue module declare from Integrations
This is weird, and I don't understand it completely. It was present
in src/pages/Integrations.vue, but of course that file is gone now.
And the frontend doesn't compile properly without it. I guess it
makes as much sense to have it here as it did to have it in
src/pages/Integrations.vue.

Signed-off-by: Adam Pickering <adamkpickering@gmail.com>