networking/mac-ca: rewrite tests.
This converts the test to just test `getMacCertificates` so that we do not
need to export the individual helper functions.  This means we no longer do
any dependency injection (we just mock `spawnFile` directly).

We also drop the test that rejected non-self-signed CAs; this can be the
case when we're chaining to a new CA (in the say that Let's Encrypt chained
to the DST root).

Signed-off-by: Mark Yen <mark.yen@suse.com>