Lima: Actually suppress sudo prompt.
This makes the checkbox in the sudo explanation window actually set a flag
to never prompt for sudo access again.

Currently, the only way to reset this flag in the UI is to do a factory
reset; as a workaround, the user can use the `rdctl` command line.

This is broken because we never write the config to disk.

Signed-off-by: Mark Yen <mark.yen@suse.com>