Support API passwords that contain embedded colons.
Not an existing bug. Right now the password is hardwired to a string
of 16 alphanumeric characters, but chances are we'll allow users to
set the password, and we need to allow it to contain ':' characters.

Bug caused by the fact that in JS

`a:b:c`.split(':', 2)` => `['a', 'b']` and the rest is discarded.

Signed-off-by: Eric Promislow <epromislow@suse.com>